+++ ISL.ReIdentification.WebApps.Configurations/ISL.ReIdentification.Configurations.Server.Tests.Integration/Apis/Audits/AuditApiTests.cs
@@ -2,13 +2,11 @@
 // Copyright (c) North East London ICB. All rights reserved.
 // ---------------------------------------------------------
 
-using FluentAssertions;
 using ISL.ReIdentification.Configurations.Server.Tests.Integration.Brokers;
 using ISL.ReIdentification.Configurations.Server.Tests.Integration.Models.Audit;
 using ISL.ReIdentification.Configurations.Server.Tests.Integration.ReIdentification.Brokers;
-using Tynamix.ObjectFiller;
 
-namespace ISL.ReIdentification.Configuration.Server.Tests.Integration.Apis
+namespace ISL.ReIdentification.Configuration.Server.Tests.Integration.Apis.Audits
 {
     [Collection(nameof(ApiTestCollection))]
     public partial class AuditsApiTests
