+++ ISL.ReIdentification.WebApps.Configurations/ISL.ReIdentification.Configurations.Server.Tests.Integration/Apis/PdsDatas/PdsDataApiTests.Delete.cs
@@ -2,25 +2,26 @@
 // Copyright (c) North East London ICB. All rights reserved.
 // ---------------------------------------------------------
 
-using FluentAssertions;
 using ISL.ReIdentification.Configurations.Server.Tests.Integration.Models.PdsData;
 
-namespace ISL.ReIdentification.Configurations.Server.Tests.Integration.Apis
+namespace ISL.ReIdentification.Configurations.Server.Tests.Integration.Apis.PdsDatas
 {
     public partial class PdsDataApiTests
     {
         [Fact]
-        public async Task ShouldGetPdsDataByIdAsync()
+        public async Task ShouldDeletePdsDataAsync()
         {
             // given
             PdsData randomPdsData = await PostRandomPdsDataAsync();
 
             // when
-            PdsData retrievedPdsData = await this.apiBroker.GetPdsDataByIdAsync(randomPdsData.Id);
+            await this.apiBroker.DeletePdsDataByIdAsync(randomPdsData.Id);
 
             // then
-            retrievedPdsData.Should().BeEquivalentTo(randomPdsData);
-            await this.apiBroker.DeletePdsDataByIdAsync(randomPdsData.Id);
+            ValueTask<PdsData> getPdsDataByIdTask =
+                this.apiBroker.GetPdsDataByIdAsync(randomPdsData.Id);
+
+            await Assert.ThrowsAsync<HttpResponseNotFoundException>(getPdsDataByIdTask.AsTask);
         }
     }
 }
