+++ ISL.ReIdentification.WebApps.Configurations/ISL.ReIdentification.Configurations.Server.Tests.Integration/Apis/PdsDatas/PdsDataApiTests.Post.cs
@@ -2,10 +2,9 @@
 // Copyright (c) North East London ICB. All rights reserved.
 // ---------------------------------------------------------
 
-using FluentAssertions;
 using ISL.ReIdentification.Configurations.Server.Tests.Integration.Models.PdsData;
 
-namespace ISL.ReIdentification.Configurations.Server.Tests.Integration.Apis
+namespace ISL.ReIdentification.Configurations.Server.Tests.Integration.Apis.PdsDatas
 {
     public partial class PdsDataApiTests
     {
