+++ ISL.ReIdentification.WebApps.Configurations/ISL.ReIdentification.Configurations.Server.Tests.Integration/Apis/PdsDatas/PdsDataApiTests.cs
@@ -2,13 +2,11 @@
 // Copyright (c) North East London ICB. All rights reserved.
 // ---------------------------------------------------------
 
-using FluentAssertions;
 using ISL.ReIdentification.Configurations.Server.Tests.Integration.Brokers;
 using ISL.ReIdentification.Configurations.Server.Tests.Integration.Models.PdsData;
 using ISL.ReIdentification.Configurations.Server.Tests.Integration.ReIdentification.Brokers;
-using Tynamix.ObjectFiller;
 
-namespace ISL.ReIdentification.Configurations.Server.Tests.Integration.Apis
+namespace ISL.ReIdentification.Configurations.Server.Tests.Integration.Apis.PdsDatas
 {
     [Collection(nameof(ApiTestCollection))]
     public partial class PdsDataApiTests
