+++ ISL.ReIdentification.WebApps.Configurations/ISL.ReIdentification.Configurations.Server.Tests.Integration/Apis/UserAccesses/UserAccessApiTests.Delete.cs
@@ -2,26 +2,24 @@
 // Copyright (c) North East London ICB. All rights reserved.
 // ---------------------------------------------------------
 
-using FluentAssertions;
 using ISL.ReIdentification.Configurations.Server.Tests.Integration.Models.UserAccess;
 
-namespace ISL.ReIdentification.Configuration.Server.Tests.Integration.Apis
+namespace ISL.ReIdentification.Configuration.Server.Tests.Integration.Apis.UserAccesses
 {
     public partial class UserAccessApiTests
     {
         [Fact]
-        public async Task ShouldGetUserAccessAsync()
+        public async Task ShouldDeleteUserAccessAsync()
         {
             // given
             UserAccess randomUserAccess = await PostRandomUserAccess();
-            UserAccess expectedUserAccess = randomUserAccess;
 
             // when
-            UserAccess actualUserAccess = await this.apiBroker.GetUserAccessByIdAsync(randomUserAccess.Id);
+            await this.apiBroker.DeleteUserAccessByIdAsync(randomUserAccess.Id);
 
             // then
-            actualUserAccess.Should().BeEquivalentTo(expectedUserAccess);
-            await this.apiBroker.DeleteUserAccessByIdAsync(actualUserAccess.Id);
+            ValueTask<UserAccess> getUserAccessByIdTask = this.apiBroker.GetUserAccessByIdAsync(randomUserAccess.Id);
+            await Assert.ThrowsAsync<HttpResponseNotFoundException>(getUserAccessByIdTask.AsTask);
         }
     }
 }
