+++ ISL.ReIdentification.WebApps.Configurations/ISL.ReIdentification.Configurations.Server.Tests.Integration/Apis/UserAccesses/UserAccessApiTests.Get.cs
@@ -2,10 +2,9 @@
 // Copyright (c) North East London ICB. All rights reserved.
 // ---------------------------------------------------------
 
-using FluentAssertions;
 using ISL.ReIdentification.Configurations.Server.Tests.Integration.Models.UserAccess;
 
-namespace ISL.ReIdentification.Configuration.Server.Tests.Integration.Apis
+namespace ISL.ReIdentification.Configuration.Server.Tests.Integration.Apis.UserAccesses
 {
     public partial class UserAccessApiTests
     {
