+++ ISL.ReIdentification.WebApps.Configurations/ISL.ReIdentification.Configurations.Server.Tests.Integration/Apis/UserAccesses/UserAccessApiTests.cs
@@ -5,9 +5,8 @@
 using ISL.ReIdentification.Configurations.Server.Tests.Integration.Brokers;
 using ISL.ReIdentification.Configurations.Server.Tests.Integration.Models.UserAccess;
 using ISL.ReIdentification.Configurations.Server.Tests.Integration.ReIdentification.Brokers;
-using Tynamix.ObjectFiller;
 
-namespace ISL.ReIdentification.Configuration.Server.Tests.Integration.Apis
+namespace ISL.ReIdentification.Configuration.Server.Tests.Integration.Apis.UserAccesses
 {
     [Collection(nameof(ApiTestCollection))]
     public partial class UserAccessApiTests
